appviewer/requirements.txt
dash 
dash-bootstrap-components 
pyelftools 
psutil
pandas
